pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "digity"
version = "0.2.2"
description = "Python SDK for the Digity sensorized exohand"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "Digity", email = "cristobal.corral@digity.de" }]
keywords = ["exohand", "glove", "sensors", "robotics", "haptics", "wearable"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
    "pyzmq>=25.0",
]

[project.optional-dependencies]
viz = [
    "flask>=3.0",
    "flask-socketio>=5.3",
    "flask-login>=0.6",
    "pywebview>=5.0",
    "werkzeug>=3.0",
]
agent = ["python-socketio[client]>=5.0", "websocket-client>=1.6"]

[project.scripts]
digity-viz   = "digity.viz:main"
digity-agent = "digity.viz:main_agent"

[project.urls]
Homepage = "https://digity.de"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"digity.viz" = ["templates/**/*", "static/**/*"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
